refactor(ConfirmUpdateDialog): move to settings/_components
